Twitterのアカウント情報取得は複数コンポーネントで使用されるのでミックスインにする
