+++ src/public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=c4f6280a950ee84d839b",
+    "/js/app.js": "/js/app.js?id=ced571208d4096f90a08",
     "/css/app.css": "/css/app.css?id=421c4c1cc2b248e6bf0d"
 }
